DRI2: Check bo dimensions before reusing
With DRI2INFOREC_VERSION >= 6 we use ARMSOCDRI2ReuseBufferNotify to
determine when to reuse a back buffer, instead of bumping the serial
number in ARMSOCDRI2CreateBuffer.

However ARMSOCDRI2ReuseBufferNotify was only checking that both the old
buffer and the new buffer were both flippable or not-flippable. This is
not correct - just because the old buffer was flipped it doesn't mean it
can be now - the size may have changed. We therefore also need to verify
the buffer size matches the drawable.

Change-Id: Ie86537d3d21a8cfe58017ef4cf02f3292174cb50